26/02/09 정재은 일감유형 모달수정 로그기능

+++ src/main/java/com/yedam/app/projectmodal/service/ProjectModalVO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ProjectModalVO {
-	private Long projectCode;
+	private Integer projectCode;
 	private String projectName;
 }
